Проект Castorama заменил parser_job
