--- parser_job/runner.py
@@ -1,14 +1,0 @@
-from twisted.internet import reactor
-from scrapy.crawler import CrawlerRunner
-from scrapy.utils.log import configure_logging
-from scrapy.utils.project import get_project_settings
-
-from spiders.labirint_ru import LabirintRuSpider
-print()
-if __name__ == '__main__':
-    print(1)
-    configure_logging()
-    settings = get_project_settings()
-    runner = CrawlerRunner(settings)
-    runner.crawl(LabirintRuSpider)
-    reactor.run()
