+++ scrapy.cfg
@@ -4,8 +4,8 @@
 # https://scrapyd.readthedocs.io/en/latest/deploy.html
 
 [settings]
-default = parser_job.settings
+default = settings
 
 [deploy]
 #url = http://localhost:6800/
-project = parser_job
+project = castorama
